move meta loading to meta report
